% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\docType{data}
\name{flags}
\alias{flags}
\title{Flags icon links}
\format{
An object of class \code{character} of length 2.
}
\usage{
flags
}
\description{
Urls of svg files for flags.
}
\details{
At this day only english ('en') and french ('fr') are set. All these icons 
are on the following website : 
\url{https://cdn.rawgit.com/lipis/flag-icon-css/master/}
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
\keyword{datasets}
