% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\docType{data}
\name{i18n}
\alias{i18n}
\title{Translation}
\format{
An object of class \code{Translator} (inherits from \code{R6}) of length 14.
}
\usage{
i18n
}
\description{
Environment set by the shiny.i18n package. Run on the translatio.json file.
}
\details{
At this day only english ('en') and french ('fr') are set in the json file. 
You can contribute to this via the github repository of this project at
\url{https://github.com/gowachin/shinyDiveR}
}
\examples{
data(i18n)
i18n$set_translation_language('fr') # choose the french language
}
\author{
Jaunatre Maxime <maxime.jaunatre@yahoo.fr>
}
\keyword{datasets}
