% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_advparamInput.R
\name{mod_advparamInput}
\alias{mod_advparamInput}
\title{mod_advparamInput}
\usage{
mod_advparamInput(id, i18n)
}
\arguments{
\item{id}{shiny id}

\item{i18n}{traduction language}
}
\description{
Shiny module for advanced parameters of a dive
}
\examples{
library(shiny)
library(shiny.i18n)
library(shinyTime)
library(shinyWidgets)
if (interactive()){
i18n <- Translator$new(translation_json_path = "inst/app/translations/translation.json")
i18n$set_translation_language('en')
ui <- fluidPage(
  mod_advparamInput('adv', i18n)
)
server <- function(input, output, session) {
}
shinyApp(ui, server)
}

}
