% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_consoUI.R
\name{mod_consoUI}
\alias{mod_consoUI}
\title{mod_consoUI}
\usage{
mod_consoUI(id, i18n, sec_plot)
}
\arguments{
\item{id}{shiny id}

\item{i18n}{traduction language}
}
\description{
Shiny module for conso
}
