% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pedaconso.R
\name{mod_peda_conso_ui}
\alias{mod_peda_conso_ui}
\title{peda_conso_ui}
\usage{
mod_peda_conso_ui(id, press, vol, maxt)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{press}{show pressure calculus}

\item{vol}{show volume calculus}
}
\description{
A shiny Module.
}
