% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_exampleUI.R
\name{mod_select_view_ui}
\alias{mod_select_view_ui}
\alias{mod_select_view_server}
\title{mod_select_view_ui and mod_select_view_server}
\usage{
mod_select_view_ui(id)

mod_select_view_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
