% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_squareUI.R
\name{mod_squareUI}
\alias{mod_squareUI}
\alias{mod_squareServer}
\alias{mod_summarisediveServer}
\title{mod_squareUI}
\usage{
mod_squareUI(id, i18n)

mod_squareServer(id, i18n)

mod_summarisediveServer(id, i18n, dive)
}
\arguments{
\item{id}{shiny id}

\item{i18n}{traduction language}

\item{dive}{a dive object created by the mn90 package}
}
\description{
Shiny module for square profile

Shiny module for square profile

Shiny module for square profile
}
\examples{
library(shiny)
library(shiny.i18n)
library(lubridate)
library(mn90)
library(shinyWidgets)
if (interactive()){
i18n <- Translator$new(translation_json_path = "inst/app/translations/translation.json")
i18n$set_translation_language('en')
ui <- fluidPage(
  mod_squareUI('square', i18n)
)
server <- function(input, output, session) {
observe({
   mod_squareServer('square', i18n)
})
}
shinyApp(ui, server)
}

}
