% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_consoUI.R, R/mod_curveUI.R,
%   R/mod_squareUI.R, R/mod_summarisedive.R
\name{mod_consoServer}
\alias{mod_consoServer}
\alias{mod_curveServer}
\alias{mod_squareUI}
\alias{mod_squareServer}
\alias{mod_summarisediveServer}
\title{mod_consoServer}
\usage{
mod_consoServer(id, i18n, sec_plot, dives, r)

mod_curveServer(id, i18n)

mod_squareUI(id, i18n)

mod_squareServer(id, i18n, r)

mod_summarisediveServer(id, i18n, dive)
}
\arguments{
\item{id}{shiny id}

\item{i18n}{traduction language}

\item{sec_plot}{input of square module if there is a second dive or not}

\item{dives}{a dive object created by the square module}

\item{r}{list passing objects from module to modules}

\item{dive}{a dive object created by the mn90 package}
}
\description{
Shiny module for square profile

Shiny module for square profile

Shiny module for square profile

Shiny module for square profile

Shiny module for square profile
}
\examples{
library(shiny)
library(shiny.i18n)
library(lubridate)
library(DiveR)
library(shinyWidgets)
if (interactive()){
i18n <- shinyDiveR::i18n
i18n$set_translation_language('en')
ui <- fluidPage(
  mod_squareUI('square', i18n)
)
server <- function(input, output, session) {
observe({
   mod_squareServer('square', i18n)
})
}
shinyApp(ui, server)
}

}
