% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(app_lang = c("en", "fr"), app_maxd = 65)
}
\arguments{
\item{app_lang}{the language for the app. 'en' by default.}

\item{app_maxd}{maximum depth. Maximum and default value is 65.}
}
\description{
Run the Shiny Application
}
